requirements.txt
streamlit>=1.39
pandas>=2.2
numpy>=1.26
plotly>=5.24
geopandas>=1.0.1
pyproj>=3.6
shapely>=2.0
python-dateutil>=2.9.0.post0
unidecode>=1.3.8